restructured server code
